requirements.txt
pytest~=8.2.2
libclang~=18.1.1
openai~=1.35.7
pyaml~=24.4.0
ninja~=1.11.1.1
meson~=1.5.1
lcov~=1.15.5a0
gcovr~=7.2
# non-python:
# python3.12 https://ubuntuhandbook.org/index.php/2023/05/install-python-3-12-ubuntu/
# python3-pip (apt) https://github.com/pypa/get-pip 
# doxygen (apt)
# graphviz (apt, dnf)
# doxygen-awesome (git submodule update --init)
# cmake (apt)
# cmocka (libcmocka-dev per apt)
# optional: some llm instance (e. g. ollama)